Fix regex error in GetHeaderGuardCPPVariable on Windows
Inserting os.sep into the expression on a Windows machine results
in an invalid expression (i.e. the escape character '\').

Since FileInfo.RepositoryName() returns a path with all forward
slashes, the correct solution is to always use '/' in the
expression, regardless of operating system.

This also fixes 18 unittest errors on Windows.