Consider foreach a known loop structure
Without this using C++11's foreach triggers a false positive
about extra space after foreach since linter considers this
a function call.

This is a bit controversial, and perhaps should be moved to
a filter somehow since in theory it might conflict with
pre-C++11 code which might have foreach declares as a regular
function.

Hope to find some solution which will satisfy everyone :)