Settings for IntelliJ IDEA Ultimate 2016
The current 5 year old settings file has several issues. E.g.:
1) The ident options for Java are not imported correctly. They still remain the default settings after import (most likely due to missing codeStyleSettings block for java language).
2) Import statement settings differ completely from written import statement order and using of blank lines (just two blocks, no empty lines within these blocks).

I have imported the settings for Eclipse in IntelliJ and adjusted some settings to fit the written style guide. This may not cover all issues but at least there is no longer a massive change of the file when you format the file in Eclipse and afterwards in IntelliJ IDEA Ultimate 2016.