[Fix] Line indentation style check for more than 4 leading spaces
The line indentation style check in the Check Style function did not check for valid leading spaces after the first 4 leading spaces. 
Resolved affected unit tests due to checking all leading spaces.
Used a variation of the included tests used by Rob Pike for line continuation testing. Assumed style guideline of only allowing matmatical operators to be used as line continuation at the end fo the previous line (+ operator is always at the end of a line so that variables will always line up).

Note: The leading spaces used for alignment in the python multi-line strings (""") in the unit tests are included in the string and become a part of the indentation testing.