(Eclipse formatter) Do not format html in Javadoc
Make eclipse formatter comply with section '7.1.2 Paragraphs' by keeping
the first word in the paragraph immediately following the paragraph tag.

Normally, Eclipse would insert a newline in the javadoc, and begin the
paragraph on the next line. That's because surrounding whitespace should
not matter for HTML block level elements like the paragraph tag, and
this improves readability of the Javadoc source. However, the Google
StyleGuide for Java forbids this extra newline in section 7.1.2. The
only way to have Eclipse respect this is by disabling html formatting
entirely.