Remove redundant clause regarding operator spaces
This clause is redundant with the first sentence in this paragraph, and makes this sentence confusing.  A hasty reader might mistake the "always be consistent" as a statement about arithmetic operators within a particular project/file.  The true intent of the clause appears to be in differentiating arithmetic operators from binary ones; but I believe that can be done more explicitly and clearly by simply listing them, as is done for the binary operators.  Further, there is no actual need to restate that consistency is necessary for binary operators since the first sentence of the paragraph puts it in black and white, and the following example shows clearly that there is no grey area.