[cpplint] detection of broken std::regex in libstdc++
Added detection of broken std::regex and friends in older
versions of libstdc++ (< 4.9) into the cpplint tool.  For details,
see [1].

The motivation for this change is the fact that some projects still
need to be compiled and linked against older libstdc++ at still
supported Linux distros such as RHEL/CentOS 7, Ubuntu 14.04 and SLES 12,
but using C++11 std::regex functionality could lead to unexpected crash
if the code is compiled with C++11-compatible compiler and linked aginst
libstdc++ of versions prior to 4.9.

References:
  [1] https://gcc.gnu.org/onlinedocs/gcc-4.8.2/libstdc++/manual/manual/status.html#status.iso.2011
      section 28 'Regular Expressions'