angularjs: fix init logic
The run_prettify.js link hasn't worked for years, so updae it to the
current cdn we use in other pages.

The styleguide.js link has always been broken.  Fixing that requires
tweaking how the TOC works.  We drop the static one in favor of the
dynamic (like other styleguides), and drop the custom <base> setting
so all links open in the current tab (like other styleguides).